feat(arc-1798): move check to item detail page
